Multimorbidity
Two do-files. These are exploratory, prior to some proper analyses in that direction. The first (cluster analysis) does a quick and dirty cluster analysis by ethnicity, age and sex. The second (variable selection) goes the other way and uses the outcome relationship (in a multivariable logistic regression) to remove pairwise interactions not of interest.